Separate board splitting and data functions
